feat: basic browser extension support
o9k comes with a UI and needs control over the running application.
These tasks are both hard to implement as classic web applications.

This patch introduces a basic manifest to support using o9k as a browser
extension. It currently does not come with an icon and simply attaches
itself to all applications running on `localhost`. This will be improved
in follow up commits.

Further down the road we might switch from a basic browser extension to
a debugger extension. This will be semantically better and gives even
more access to some APIs we might be interested.